refactor(server): split monolithic index.ts into focused modules
Refactor 3,183-line index.ts into 13 maintainable modules organized by responsibility:
- config/godot-path.ts: Godot executable detection and validation
- operations/: 4 modules handling project, scene, UID, and LSP operations
- tools/: Tool registry and routing
- utils/: Shared validation and parameter conversion utilities

Entry point reduced to 18 lines. All 19 MCP tools preserved with no breaking changes.
